refactor: moved error_handler to resources
